+++ composer.json
@@ -25,7 +25,7 @@
             "src/Tightenco/Support/alias.php"
         ],
         "psr-4": {
-            "Tightenco\\": "src/Tightenco"
+            "Tightenco\\Collect\\": "src/Tightenco"
         }
     },
     "scripts": {
