+++ src/Tightenco/Contracts/Support/Arrayable.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tightenco\Contracts\Support;
+namespace Tightenco\Collect\Contracts\Support;
 
 interface Arrayable
 {
